Reduce camera scale and grid size in ryot_compass
This commit reduces the maximum camera scale from 50.0 to 5.0 in the helper function 'camera.rs'. It also reduces the grid size in 'main.rs', going from a size of 1100x1100 to 200x120, removing the condition to only create elements when coordinates are multiple of 10. This guarantees a maximum zoom out of 20%, ensuring a smoother and more manageable user experience with less computational load.