feat: Improve general usage of cip sprites within bevy.
- Decompressed sheets cache
We decompress and save all sprite-sheets into disk. This makes loading of individual sprites 50x faster in detriment of 2Gb of disk space.
Decompression only happens if the decompressed version is not found. In the future we can turn decompression into a button/action, that can decompress multiple sheets.

- Gradual InMemory sprite cache
The first time a sprite is loaded it's cached in memory for faster track. This can occupy up to 6Gb of memory, if all sprites are loaded. However, it makes loading 1000x faster (1s for every 10k sprites vs 1ms for every 10k sprites). We can balance this in the future with optional caching and custom cache TTL.