feat: Change drawing details as we zoom in/out
We want to be able to reduce/increase the amount of visible entities as we zoom in and out. This is good to keep the performance decent while showing bigger areas of the map.

This also allows to have a better view of the map silhouette when in very zoomed out views.