feat: Introduce the concept of BrushParams
This allows us to have different ways of interacting with the brush. Sometimes we want to input by sliding/holding, but sometimes we just want the brush to be inserted in two clicks (to define a row or a line or an irregular shape).

This makes the brushes more flexible as we evolve.