impr: Old/new info to always target the same layer
In a same UpdateTileContent command, the old and new info should always
target the same layer. They shouldn't ever point to different layers.

In a bottom layer, if we are drawing in the next slot in the stack, the
old info should point to this slot with an empty appearance, because a
revert in this case is just a deletion of that layer.