feat: Game Interactions
This commit introduces the concept of Game Interactions, a pipeline of systems that filters out positions from a list given the existence or not of a group of Components in the entities contained in the relevant positions.

It is useful for performing things like collision check, visibility validation or others.