impr: re-vamp nomenclature for perspective module
We are polishing the nomenclature of our perspective module.

ViewPoint is now Traversal, closer to the representation of an area that can be traversed by a ray cast.

ConditionalViewPointApp is now Trajectory, closer to the representation of a connection between two points, given certain conditions.

Sight and Path are now respectively VisibleTrajectory and WalkableTrajectory.